+++ src/canary.ts
@@ -151,6 +151,10 @@ export type { ArdaSelectProps } from './components/canary/atoms/select';
 export { AutoFillLabel } from './components/canary/atoms/auto-fill-label';
 export type { AutoFillLabelProps } from './components/canary/atoms/auto-fill-label';
 
+// Molecules: AddressFieldset
+export { AddressFieldset } from './components/canary/molecules/address-fieldset';
+export type { AddressFieldsetProps } from './components/canary/molecules/address-fieldset';
+
 // Molecules: AutoFillField
 export { AutoFillField } from './components/canary/molecules/auto-fill-field';
 export type { AutoFillFieldProps } from './components/canary/molecules/auto-fill-field';